requirements.txt
# LLM 및 API 관련
openai>=1.0.0
python-dotenv>=1.0.0

# 템플릿 엔진
jinja2>=3.0.0

# 진행률 표시
tqdm>=4.60.0

# 데이터 처리
pandas>=1.5.0
numpy>=1.21.0

# 병렬 처리
concurrent-futures>=3.1.0

# 파일 압축 해제 (xz 파일 처리용)
lzma>=3.0.0

# 과학 계산 (선택사항)
scipy>=1.9.0
scikit-learn>=1.1.0

# 시각화
matplotlib>=3.5.0
seaborn>=0.12.0

# MCP 서버 (Model Context Protocol)
mcp>=1.0.0 